基本块合并优化、乘除指令优化

+++ backend/instructions/MemInst/LI.java
@@ -5,9 +5,9 @@ import backend.instructions.MIPSInst;
 
 public class LI extends MIPSInst {
     public Register target;
-    public int immediate;
+    public long immediate;
 
-    public LI(Register target, int immediate) {
+    public LI(Register target, long immediate) {
         this.target = target;
         this.immediate = immediate;
     }
